# docs/swagger.yaml
basePath: /
definitions:
...
  CurrencyControl_internal_delivery_dto.TrashItem:
    properties:
      amount:
        type: number
      client_id:
        type: integer
      client_name:
        type: string
      contract_id:
        type: integer
      contract_number:
        type: string
      created_by:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      document_date:
        type: string
      document_path:
        type: string
      entity_name:
        type: string
      entity_type:
        type: string
      id:
        type: integer
      number:
        type: string
    type: object
  CurrencyControl_internal_delivery_dto.TrashListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/CurrencyControl_internal_delivery_dto.TrashItem'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  delivery_http.CommonError:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: API сервера для системы валютного контроля.
  title: Currency Control API
  version: "1.0"
paths:
...
  /api/trash:
    get:
      description: Возвращает список всех удаленных документов (контракты, доп. соглашения,
        инвойсы, ГТД). Доступно только сотрудникам Комплаенс и Валютного контроля.
      parameters:
      - description: Фильтр по типу (contract, additional_agreement, invoice, gtd)
        in: query
        name: entity_type
        type: string
      - description: Фильтр по ID филиала
        in: query
        name: branch_id
        type: integer
      - description: Поиск по номеру документа или контрагенту
        in: query
        name: search
        type: string
      - description: Номер страницы (по умолчанию 1)
        in: query
        name: page
        type: integer
      - description: Размер страницы (по умолчанию 20, макс 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CurrencyControl_internal_delivery_dto.TrashListResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Просмотр корзины удаленных документов
      tags:
      - Trash
  /api/trash/{entity_type}/{id}:
    get:
      description: Возвращает подробную карточку удаленного документа из корзины.
        Доступно только Комплаенс и Валютному контролю.
      parameters:
      - description: Тип сущности (contract, additional_agreement, invoice, gtd)
        in: path
        name: entity_type
        required: true
        type: string
      - description: ID записи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CurrencyControl_internal_delivery_dto.TrashItem'
        "400":
          description: Некорректные параметры
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "404":
          description: Документ не найден
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Детальная информация об удаленном документе
      tags:
      - Trash
  /api/trash/{entity_type}/{id}/file:
    get:
      description: Отдает файл удаленного документа (PDF/Word) для просмотра. Доступно
        только Комплаенс и Валютному контролю.
      parameters:
      - description: Тип сущности (contract, additional_agreement, invoice, gtd)
        in: path
        name: entity_type
        required: true
        type: string
      - description: ID записи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Файл документа
          schema:
            type: file
        "400":
          description: Некорректные параметры
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "404":
          description: Файл не найден
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Просмотр или скачивание удаленного файла из корзины
      tags:
      - Trash
  /api/trash/{entity_type}/{id}/restore:
    post:
      description: Восстанавливает удаленный документ обратно в систему. Доступно
        только Комплаенс и Валютному контролю.
      parameters:
      - description: Тип сущности (contract, additional_agreement, invoice, gtd)
        in: path
        name: entity_type
        required: true
        type: string
      - description: ID записи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном восстановлении
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректные параметры
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Восстановление документа из корзины
      tags:
      - Trash
...
securityDefinitions:
  ApiKeyAuth:
    description: Введите Bearer access_token (полученный при /auth/login или /auth/refresh)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"